requirements.txt
asgiref==3.3.4
autopep8==1.5.6
Django==3.2
django-filter==2.4.0
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
fontawesome-free==5.15.3
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
